requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
pydantic-settings==2.9.1
httpx==0.28.1
google-cloud-aiplatform==1.88.0
google-cloud-discoveryengine==0.13.7
google-auth==2.40.1
python-dotenv==1.1.0
